Feature/add org environment file (#123)
This PR adds a file environments.org.ts file to the .gitignore file.
This file will be referenced with the new way to execute the application
locally when someone is a part of the organization. We do this by adding
another configuration to the angular.json file for an "org" environment.
In addition, two new scripts were added to the package.json file to work
with this.