feat: Use local instead of org and auto add db to pgadmin (#132)
I added a pgpass and servers.json file to the .docker folder to be used
in the docker-compose.yml file. These allow us to setup the pgadmin
container to automatically add an "Involvemint Local" server with access
to the postgres container so that new engineers do not have to worry
about adding the server or authenticating with the local involvemin
database in order to view it from pgadmin.

Updated the init.sql scrip to successfully create the involvemin
database when first running the docker compose up command. The previous
script failed due to the "if not exists" part.

Updated the build process locally to use the local flag instead of the
org flag. This is more logical from an engineers perspective. It means
that the environment file will be environment.local.ts and the commands
to run the server and client are npm run start:server:local and npm run
start:client:local respectively.

I updated the readme to reflect the changes made in this branch.