Update Postgres Docker Container Host Port (#133)
Updated the docker compose to use 5433 as the locally accessible port in
order to not clash with any locally running postgres instances outside
of the docker container. We were running into issues where if someone
already has a local postgres server running, the application was
connecting to it instead of the docker container. This is fine as long
as the developer knows where to look, but for getting DFG running as
quick as possible this eliminates a bit of confusion.