Feature/add firebase emulator container (#134)
Updates to use a default address when running locally instead of
depending on access to the geolocation api. The default address ensures
that the default latitude and longitude match the addresses.

Adds docker container that host the firebase emulator suite so that we
can run the application without using a gcp storage bucket.